==> 4_sa.ipynb <==
#1
%matplotlib inline
import pandas as pd
import pickle
import numpy as np
import matplotlib.pyplot as plt
from consts import sensor_day_seconds as day_second
from consts import sensor_idd as idd
from consts import sensor_idds_uniq as idds_uniq
from consts import sensor_idds_uniq_sorted as idds_uniq_sorted
from consts import sensor_a_temperature as a_temperature
from consts import sensor_s_temperature as s_temperature
from consts import sensor_unix_column as unix_column
from consts import sensor_idd as idd
epoch_length = 30
import multiprocessing as mp
from multiprocessing import Pool
import time

#2
input_file_new = open('3_dict_list_new.pkl','rb')
input_list_new = pickle.load(input_file_new)
input_file_new.close()

#3
# 将epoch epoch_len epoch_time 都flat化
len_list_new = []
time_stamp_list_new = []
epoch_list_new = []
for start_time in input_list_new:
    # print(len(start_time[0]))
    for dict in start_time[0]:
        len_list_new.append(len(dict))
        epoch_list_new.append(dict)
    start_point = start_time[1]
    time_stamps = np.arange(start_point,start_point+day_second+epoch_length,epoch_length)
    time_stamp_list_new.extend(time_stamps[:-1])
len_array_new = np.array(len_list_new)
time_stamp_array_new = np.array(time_stamp_list_new) # 时间序列
epoch_array_new = np.array(epoch_list_new)

start_list = np.arange(1162393768,1178747998+day_second,day_second)
ddf_list_new = []
for start_point in start_list:
    time_stamps = np.arange(start_point,start_point+day_second+epoch_length,epoch_length)
    sub_df_list = []
    df = pd.read_hdf('day2.h5',str(start_point))
    for i in range(len(time_stamps)-1):
        sub_df = df[(df[unix_column] > time_stamps[i]) & (df[unix_column] <= time_stamps[i+1])]
        sub_df_list.append(sub_df)
    ddf_list_new.extend(sub_df_list)

# 修改为多线程化读入

#4
def read_sub_df(start_point):
    time_stamps = np.arange(start_point,start_point+day_second+epoch_length,epoch_length)
    sub_df_list = []
    df = pd.read_hdf('day2_sa.h5',str(start_point))
    for i in range(len(time_stamps)-1):
        sub_df = df[(df[unix_column] > time_stamps[i]) & (df[unix_column] <= time_stamps[i+1])]
        
        sub_df_list.append(sub_df)
    print("Done \t {}".format(start_point))
    return (start_point,sub_df_list)

#5
start_list = np.arange(1162393768,1178747998+day_second,day_second)
manager = mp.Manager()
# aggreateData = manager.dict()
p = Pool(16)
tmp_result = []

for start_point in start_list:
    result = p.apply_async(func=read_sub_df,args=(start_point,))
    tmp_result.append(result)
p.close()
p.join()


#6
origin_time = start_list[0]

results = []
for i in start_list:
    results.append([])

for item in tmp_result:
    temp = item.get()
    results[int((temp[0]-origin_time)/86400)].extend(temp[1])

ddf_list_new = []
for item in results:
    ddf_list_new.extend(item)

#7
index1 = np.where(len_array_new > 40)[0]
maybe_array_new = epoch_array_new[index1]
print(len(maybe_array_new))

#8
# init maybe_dict 
maybe_dict = {}
for i in idds_uniq:
    maybe_dict[i] = 0


for item in maybe_array_new:
    uniq_item = np.unique(item)
    for j in uniq_item:
        maybe_dict[j] += 1
print(maybe_dict)

#9                                                      
maybe_dict_sorted = sorted(maybe_dict.items(),key=lambda kv:(kv[1],kv[0]),reverse=True)
maybe_dict_sorted = np.array(maybe_dict_sorted)

sub_sensor_list = maybe_dict_sorted.T[0]
print(sub_sensor_list)

#10
# 测试sensor数量为多少合适
def get_sub_index(sensor_count):
    sub_sensor_list = idds_uniq_sorted[:sensor_count]
    sub_index = []
    for i, item in enumerate(epoch_array_new):
        uniq_item = np.unique(item)
        diff = np.setdiff1d(sub_sensor_list,uniq_item)
        if len(diff) == 0:
            sub_index.append(i)
    return len(sub_index),sub_index

epoch_array_new.shape

for i,item in enumerate(sub_index_array):
    print("id",i+10,"number:",item[0])

len(sub_index_array)

#11 多进程版 
start_time = time.time()
def multi_get_sub_index(i):
    return get_sub_index(i)

sensor_count_list = np.linspace(10,50,41,dtype=int)

manager = mp.Manager()
# aggreateData = manager.dict()
p = Pool(16)
tmp_result = []

for i in sensor_count_list:
    result = p.apply_async(func=multi_get_sub_index,args=(i,))
    tmp_result.append(result)
p.close()
p.join()
print(time.time()-start_time)

#11-1
sub_index_list = []
for item in tmp_result:
    sub_index_list.append(item.get())
sub_index_array = np.array(sub_index_list)

sub_index_array.shape

# obsolete 12 
def get_length_count(window):
    for item in sub_index_array:
        zeros = np.zeros_like(epoch_array)
        zeros[item[1]] = 1
        sum = 0
        for j in range(len(zeros)):
            tmp = zeros[j:j+window]
            if 0 not in tmp:
                sum += 1
        print(sum)

# 12-1 DUMP
output_file = open('sub_index_array_drop_na_sa.pkl','wb')
pickle.dump(sub_index_array,output_file)
output_file.close()

# 12-2 LOAD
input_file = open('sub_index_array_drop_na_sa.pkl','rb')
sub_index_array = pickle.load(input_file)
input_file.close()

#13
def get_data_matrixs(sensor_count,window,sub_index_array = None):
    print(idds_uniq_sorted[:sensor_count])
    tmp = sub_index_array[sensor_count-10][1]
    zeros = np.zeros_like(epoch_array_new)
    zeros[tmp] = 1

    sum_list = []
    for i in range(len(zeros)):
        tmp1 = zeros[i:i+window]
        if 0 not in tmp1:
            sum_list.append(i)
            
    return sum_list
indices = get_data_matrixs(19,7,sub_index_array)
print(len(indices))

len(indices)

indices[5]

#14
# 丢弃多余帧
def drop_useless_sensor(df,sensor_list):
    # print(sensor_list)
    for i in df[idd].values:
        if i not in sensor_list:
            df = df[df[idd] != i]
    # print(len(df))
    if len(df) < len(sensor_list):
        return None
    return df
# t = drop_useless_sensor(ddf_list[indices[3]],idds_uniq_sorted[:19])
# print(len(t))

# ignore
def handle_df(ddf_list,indice,sensor_list):
    df = ddf_list[indice]
    df = df.drop_duplicates(idd)
    df = drop_useless_sensor(df,sensor_list)
    return df

df_sample = handle_df(ddf_list,indices[0],idds_uniq_sorted[:19])

# before 15
for i, df in enumerate(ddf_list):
    with open('./data/{}.pkl'.format(i),'wb') as f:
        pickle.dump(df,f)

#15 multi_processing
def multi_convert_df_to_dm(indice,sensor_list,window=7):
    df_list = []
    for i in range(window):
        with open('./data/{}.pkl'.format(indice+i),'rb') as f:
            sub_df = pickle.load(f)
        df_list.append(sub_df)
    return_list = []
    for df in df_list:
        df = df.drop_duplicates(idd)
        df = drop_useless_sensor(df,sensor_list)
        if type(df) == type(None):
            break
        df = df.sort_values(by=idd)
        flat_matrix = np.zeros(100)
        indice_dict = {63:10,23:14,17:15,47:20,7:21,27:22,26:24,14:25,121:26,46:27,8:28,111:32,75:33,57:34,59:35,40:36,34:37,61:41,71:42,65:43,82:45,62:46,76:48,51:51,11:52,70:53,49:54,60:57,84:60,81:61,55:62,72:63,73:64,35:66,31:67,69:68,79:69,103:70,80:71,96:72,97:75,100:76,33:77,19:78,93:79,109:80,87:81,104:84,89:91,88:92,98:94}
        return_list.append(df[s_temperature].values)
    return np.array(return_list).T
res_array = multi_convert_df_to_dm(indices[46],idds_uniq_sorted[:19],7)
print(res_array)

#15
def convert_df_to_dm(ddf_list,indice,sensor_list,window=7):
    # print(sensor_list)
#     print(len(sensor_list))
    df_list = []
    for i in range(window):
        df_list.append(ddf_list[indice+i])
#     df = ddf_list[indice]
    return_list = []
    for df in df_list:
        df = df.drop_duplicates(idd)
#         print(indice,type(df))
#         print(len(df))
        df = drop_useless_sensor(df,sensor_list)
        if type(df) == type(None):
            break
        df = df.sort_values(by=idd)
        flat_matrix = np.zeros(100)
        indice_dict = {63:10,23:14,17:15,47:20,7:21,27:22,26:24,14:25,121:26,46:27,8:28,111:32,75:33,57:34,59:35,40:36,34:37,61:41,71:42,65:43,82:45,62:46,76:48,51:51,11:52,70:53,49:54,60:57,84:60,81:61,55:62,72:63,73:64,35:66,31:67,69:68,79:69,103:70,80:71,96:72,97:75,100:76,33:77,19:78,93:79,109:80,87:81,104:84,89:91,88:92,98:94}
        return_list.append(df[s_temperature].values)
#         for i,indice in enumerate(df[idd].values):
#             flat_matrix[indice_dict[indice]] = df[s_temperature].values[i]
    return np.array(return_list).T
    

res_array = convert_df_to_dm(ddf_list,indices[45],idds_uniq_sorted[:19],7)
print(res_array)

for i,df in enumerate(ddf_list):
    print(i)

res_array

#16多线程
array_list2 = []
manager = mp.Manager()

p = Pool(16)

start_time = time.time()
def multi_get_sub_index(i):
    return get_sub_index(i)

tmp_result = []

for i in indices:
    result = p.apply_async(func=multi_convert_df_to_dm,args=(i,idds_uniq_sorted[:19],7,))
    tmp_result.append(result)
p.close()
p.join()

for item in tmp_result:
    array_list2.append(item.get())

print(time.time()-start_time)

#16
start_time = time.time()
array_list2 = []
for i in indices:
    array_list2.append(convert_df_to_dm(ddf_list,i,idds_uniq_sorted[:19],7))
print(time.time()-start_time)

#before 17-4
sum = 0
new_indices = []
for i,item in enumerate(array_list2):
    if np.isnan(item).any():
        print(i)
#         sum+=1
    else:
        new_indices.append(i)

time_stamp_array

#17-1
with open('4_tensor.pkl','wb') as f:
    pickle.dump(array_list2,f)

#17-2
with open('4_time_stamp_array.pkl','wb') as f:
    pickle.dump(time_stamp_array,f)

#17-3
with open('4_indices_na.pkl','wb') as f:
    pickle.dump(indices,f)

#17-4
with open('4_indices_no_na.pkl','wb') as f:
    pickle.dump(new_indices,f)
