==> 5_tensor_prepare.ipynb <==
#1
%matplotlib inline
import pandas as pd
import pickle
import numpy as np
import matplotlib.pyplot as plt
from consts import sensor_day_seconds as day_second
from consts import sensor_idd as idd
from consts import sensor_idds_uniq as idds_uniq
from consts import sensor_idds_uniq_sorted as idds_uniq_sorted
from consts import sensor_a_temperature as a_temperature
from consts import sensor_s_temperature as s_temperature
from consts import sensor_unix_column as unix_column
from consts import sensor_idd as idd
epoch_length = 30

#2-1
with open('4_tensor.pkl','rb') as f:
    tensor_array = pickle.load(f)

tensor_array = np.array(tensor_array)

#2-2
with open('4_time_stamp_array.pkl','rb') as f:
    time_stamp_array = pickle.load(f)

#2-3
with open('4_indices_na.pkl','rb') as f:
    indices = pickle.load(f)

#2-4
with open('4_indices_no_na.pkl','rb') as f:
    new_indices = pickle.load(f)

#2-5
tensor_array = tensor_array[new_indices]

type(tensor_array)

len(tensor_array)

len(time_stamp_array)

## try to save tensor_array to dataloader

#3
import torch
import torch.utils.data as Data

for i,item in enumerate(tensor_array):
    if np.isnan(item).any():
        print(i)

tensor_array[9671]

#4
index = torch.tensor(np.linspace(1,len(tensor_array),len(tensor_array),dtype=int))

torch_dataset = Data.TensorDataset(torch.tensor(tensor_array,dtype=torch.float32))

#5
tmp = int(len(tensor_array)*0.9)
ratio = [tmp,len(tensor_array)-tmp]

#6
train_db,test_db = Data.random_split(torch.tensor(tensor_array,dtype=torch.float32),ratio)

#7-1
train_loader = Data.DataLoader(
    dataset=train_db,
    batch_size=500,
    shuffle=True,
    num_workers=8
)

#7-2
test_loader = Data.DataLoader(
    dataset=test_db,
    batch_size=500,
    shuffle=True,
    num_workers=8
)

#8-1
torch.save(train_loader,'5_train2.pt')

#8-2
torch.save(test_loader,'5_test2.pt')
