==> 19.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from rich.progress import track

a_temper = 'Ambient Temperature'
s_temper = 'Surface Temperature'


df_sum = pd.read_hdf('data_ori_na.h5','df')

# df_sum_sa = pd.read_hdf('r_data_files/data_sa_na.h5','df')

# time_32 = df_sum['time'].values
# time_32 = time_32/32
# time_32 = time_32.astype(int)
# df_sum['time_32'] = time_32


# df_sum.to_hdf('data_ori_na.h5','df')

# df_sum.to_hdf('r_data_files/data_ori_na.h5','df')

total_id_list = np.unique(df_sum['Station ID'].values)
print(total_id_list)

total_id_list

df_sum.sort_values(by='Station ID')

df_sum[df_sum['time_32']==1].sort_values(by='Station ID')

time_range = np.unique(df_sum['time_32'].values)
print(time_range)

import pickle

result_array = np.ones((np.max(time_range)+1,np.max(total_id_list)+1))*(-1)
for time_id in track(time_range):
    # test_array = np.ones(np.max(total_id_list)+1)*(-1)
    result_array[time_id][total_id_list] = 0
    tmp_df = df_sum[df_sum['time_32'] == time_id]
    id_list = tmp_df['Station ID'].values
    a_temper_list = tmp_df[a_temper].values
    result_array[time_id][id_list] = a_temper_list


with open('readings_timeid2.pkl','wb') as f:
    pickle.dump(result_array,f)    

for id in total_id_list:
    print(id,',',result_array[1232][id])

len(time_range)

np.max(time_range)

test_array[total_id_list]

len(np.unique(id_list))

total_id_list = np.unique(df_sum['Station ID'].values)
print(total_id_list)


test_array = np.ones(np.max(total_id_list)+1)*(-1)
print(test_array)

test_array[total_id_list] = 0

test_array[total_id_list[:4]] = 1

test_array[total_id_list]

np.where(test_array == 1)

df_sum[df_sum['time_32'] == 1232].sort_values(by='Station ID')
